requirements.txt
tiktoken==0.4.0
langchain==0.0.230
bokeh==2.4.3
contourpy==1.1.0
fpdf==1.7.2
streamlit==1.27.2
streamlit-bokeh-events==0.1.2
xyzservices==2023.5.0
streamlit-echarts==0.4.0
pyecharts==2.0.3
streamlit-aggrid==0.3.4.post3
extra-streamlit-components==0.1.56
google-search-results==2.4.2
streamlit-chat==0.1.1
openai==0.27.8
matplotlib==3.7.2
uniplot==0.10.0
pytest_asyncio==0.21.1
seaborn==0.13.0
